requirements.txt
mysql-connector==2.1.6
nose==1.3.7
setuptools
wheel
twine
